requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0

# Database and Session Storage
psycopg2-binary==2.9.7
redis==5.0.1

# Audio processing
numpy==1.24.3
scipy==1.11.2
librosa==0.10.1
sounddevice==0.4.6
pyaudio==0.2.11

# Security and validation
python-dotenv==1.0.0
cryptography==41.0.7

# WebSocket support
python-socketio==5.8.0
eventlet==0.33.3

# Utilities
requests==2.31.0
python-dateutil==2.8.2